deprecated old TextEditUtils Created org.eclipse.lemminx.commons.DiagnosticUtils, updated some tests to use these Created org.eclipse.lemminx.commons.TextEditUtils Created org.eclipse.lemminx.utils.DOMTextEditUtils
