+++ org.eclipse.lemminx/src/main/java/org/eclipse/lemminx/extensions/contentmodel/participants/codeactions/RootElementTypeMustMatchDoctypedeclCodeAction.java
@@ -15,11 +15,11 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.eclipse.lemminx.commons.CodeActionFactory;
+import org.eclipse.lemminx.commons.DiagnosticUtils;
 import org.eclipse.lemminx.dom.DOMDocument;
 import org.eclipse.lemminx.dom.DOMElement;
 import org.eclipse.lemminx.services.extensions.codeaction.ICodeActionParticipant;
 import org.eclipse.lemminx.services.extensions.codeaction.ICodeActionRequest;
-import org.eclipse.lemminx.utils.MarkupContentFactory;
 import org.eclipse.lemminx.utils.XMLPositionUtility;
 import org.eclipse.lsp4j.CodeAction;
 import org.eclipse.lsp4j.Diagnostic;
@@ -53,7 +53,7 @@ public class RootElementTypeMustMatchDoctypedeclCodeAction implements ICodeActio
 			return;
 		}
 
-		String message = MarkupContentFactory.getDiagnosticMessage(diagnostic);
+		String message = DiagnosticUtils.getDiagnosticMessage(diagnostic);
 		
 		String currentRootText = getCurrentRoot(message);
 		if (currentRootText == null || !currentRootText.equals(root.getNodeName())) {
